Extract an actual overwritten LiveCollectionTrait
