+++ src/Sylius/Bundle/AdminBundle/Twig/Component/Taxon/FormComponent.php
@@ -14,6 +14,7 @@ declare(strict_types=1);
 namespace Sylius\Bundle\AdminBundle\Twig\Component\Taxon;
 
 use Sylius\Bundle\AdminBundle\Generator\TaxonSlugGeneratorInterface;
+use Sylius\Bundle\UiBundle\Twig\Component\LiveCollectionTrait;
 use Sylius\Bundle\UiBundle\Twig\Component\ResourceFormComponentTrait;
 use Sylius\Component\Core\Model\TaxonInterface;
 use Sylius\Resource\Doctrine\Persistence\RepositoryInterface;
@@ -25,6 +26,7 @@ use Symfony\UX\LiveComponent\Attribute\LiveArg;
 #[AsLiveComponent]
 class FormComponent
 {
+    use LiveCollectionTrait;
     /** @use ResourceFormComponentTrait<TaxonInterface> */
     use ResourceFormComponentTrait;
