+++ src/Sylius/Bundle/UiBundle/Twig/Component/LiveCollectionTrait.php
@@ -1,0 +1,65 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Sylius\Bundle\UiBundle\Twig\Component;
+
+use Symfony\Component\PropertyAccess\PropertyAccessorInterface;
+use Symfony\UX\LiveComponent\Attribute\LiveAction;
+use Symfony\UX\LiveComponent\Attribute\LiveArg;
+use Symfony\UX\LiveComponent\ComponentWithFormTrait;
+
+/** @see \Symfony\UX\LiveComponent\LiveCollectionTrait */
+trait LiveCollectionTrait
+{
+    use ComponentWithFormTrait;
+
+    #[LiveAction]
+    public function addCollectionItem(
+        PropertyAccessorInterface $propertyAccessor,
+        #[LiveArg] string $name,
+        #[LiveArg] ?string $type = null,
+    ): void {
+        $propertyPath = $this->fieldNameToPropertyPath($name, $this->formName);
+        $data = $propertyAccessor->getValue($this->formValues, $propertyPath);
+
+        if (!\is_array($data)) {
+            $propertyAccessor->setValue($this->formValues, $propertyPath, []);
+            $data = [];
+        }
+
+        $index = [] !== $data ? count($data) : 0;
+        $propertyAccessor->setValue(
+            $this->formValues,
+            sprintf('%s[%s]', $propertyPath, $index),
+            null === $type ? [] : ['type' => $type],
+        );
+    }
+
+    #[LiveAction]
+    public function removeCollectionItem(
+        PropertyAccessorInterface $propertyAccessor,
+        #[LiveArg] string $name,
+        #[LiveArg] int|string $index,
+    ): void {
+        $propertyPath = $this->fieldNameToPropertyPath($name, $this->formName);
+        $data = $propertyAccessor->getValue($this->formValues, $propertyPath);
+        unset($data[$index]);
+        $propertyAccessor->setValue($this->formValues, $propertyPath, $data);
+    }
+
+    private function fieldNameToPropertyPath(string $collectionFieldName, string $rootFormName): string
+    {
+        $propertyPath = $collectionFieldName;
+
+        if (str_starts_with($collectionFieldName, $rootFormName)) {
+            $propertyPath = substr_replace($collectionFieldName, '', 0, mb_strlen($rootFormName));
+        }
+
+        if (!str_starts_with($propertyPath, '[')) {
+            $propertyPath = "[$propertyPath]";
+        }
+
+        return $propertyPath;
+    }
+}
